requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-admin-dracula==0.0.3
django-cors-headers==4.9.0
django-nested-admin==4.1.4
djangorestframework==3.16.1
idna==3.10
pillow==11.3.0
python-monkey-business==1.1.0
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
